+++ src/frontend/main.ts
@@ -1,21 +1,5 @@
 import { createApp } from "vue";
-import axios, { type AxiosInstance } from "axios";
 import App from "@/component/App.vue";
 import { router } from "@/routes";
-import { getAccessToken } from "@/service/authService";
 
-function bootstrap() {
-  createApp(App).use(router).mount("#app");
-}
-
-export const axiosInstance: AxiosInstance = axios.create({});
-
-axiosInstance.interceptors.request.use(async (config) => {
-  const token = await getAccessToken();
-  if (token) {
-    config.headers.Authorization = `Bearer ${token}`;
-  }
-  return config;
-});
-
-bootstrap();
+createApp(App).use(router).mount("#app");