+++ vite.config.js
@@ -23,4 +23,16 @@ export default defineConfig({
     emptyOutDir: true,
   },
   assetsInclude: ["**/*.cert"],
+  server: {
+    proxy: {
+      "/api": {
+        target: "http://localhost:3000",
+        changeOrigin: true,
+      },
+      "/socket.io": {
+        target: "http://localhost:3000",
+        ws: true,
+      },
+    },
+  },
 });